requirements.txt
sympy
numpy
discord.py
python-dotenv
